Refactor git component out from runReport
